Рефакторинг UIAnnotationView
